moved NCBI fetching tests from unit/ to integration/
Since the NCBI-fetching tests hit a third-party API, there is a chance
that they will fail when the server is down. The NCBI fetching
functions themselves are straightforward, so the tests are mostly
validation that we are interacting with the NCBI API in the correct way
for parameters like db name, and rettype. The tests have been moved to
integration so they will only be run during full testing runs.
